Add files via upload
This project is based on Titanic Dataset which is been extracted from Kaggle. This project is to predict how many of them survived in titanic Crash. The target variable in this is  "Survived". In this project, I have used different libraries (like numpy, pandas, matplotlib, ML models) on dataset to collect, understand, preprocess, and build the machine learning models for finding the accuracy related to actual VS predicted values.